+++ R/pairwise-comparisons.R
@@ -58,7 +58,7 @@
 #' pairwise <- pairwise_comparison(scores, by = "target_type")
 #'
 #' library(ggplot2)
-#' plot_pairwise_comparison(pairwise, type = "mean_scores_ratio") +
+#' plot(pairwise, type = "mean_scores_ratio") +
 #'   facet_wrap(~target_type)
 
 pairwise_comparison <- function(scores,
@@ -138,6 +138,8 @@ pairwise_comparison <- function(scores,
   )
 
   out <- data.table::rbindlist(results)
+
+  class(out) <- c("scoringutils_pairwise", class(out))
 
   return(out[])
 }